% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialsurvival.R
\name{expsurv.bootstrap}
\alias{expsurv.bootstrap}
\title{Bootstrap a parametric exponential mixture model}
\usage{
expsurv.bootstrap(alpha, lambda, N, duration, samples)
}
\arguments{
\item{alpha}{Proportion of individuals who experience 0 risk}

\item{lambda}{Constant hazard}

\item{N}{Number of realisations (individuals in population)}

\item{duration}{Duration of the study}

\item{samples}{Number of samples to bootstrap.

Recommended number of samples is at least \code{N}.}
}
\value{
Data frame with \code{samples} rows, with each row containing an estimate of \eqn{\alpha} and \eqn{\lambda} from a simulation.
}
\description{
Performs bootstrapping on the parametric exponential mixture model, to provide approximate measures on the uncertainty.
}
\examples{
bootstrapping <- expsurv.bootstrap(alpha = 0.3, lambda = 1, N = 1000, duration = 15, samples = 100)
}
