% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialsurvival.R
\name{expsurv.fit}
\alias{expsurv.fit}
\title{Fit Parametric Exponential Mixture Model}
\usage{
expsurv.fit(surv)
}
\arguments{
\item{surv}{A matrix with the first column being event times (or censor time), and the second being whether the event was censored. 

Note: For censoring, 0=alive (censored), 1=dead}
}
\value{
Data frame with two elements: Zero risk proportion, \eqn{\alpha}, and estimated hazard, \eqn{\lambda}.
}
\description{
Fits a parametric mixture model of the form \eqn{S(t) = \alpha + (1-\alpha)exp(-\lambdat)} to (right-censored) data.
}
\details{
This form of model allows for some proportion of individuals to experience zero risk (\eqn{\alpha}), and the remainder to experience constant hazard (\eqn{\lambda}).
}
\examples{
surv <- expsurv.simulate(alpha = 0.3, lambda = 1, N=1000, duration = 15)
fit <- expsurv.fit(surv)
}
