% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialsurvival.R
\name{expsurv.likelihood}
\alias{expsurv.likelihood}
\title{Likelihood for a Parametric Exponential Mixture Model}
\usage{
expsurv.likelihood(alpha, lambda, data)
}
\arguments{
\item{alpha}{Proportion of individuals whom experience 0 risk}

\item{lambda}{Constant hazard}

\item{data}{A matrix with the first column being event times (or censor time), and the second being whether the event was censored. 

Note: For censoring, 0=alive (censored), 1=dead}
}
\description{
Calculates the likelihood function for a Parametric Exponential Mixture Model
}
\examples{
surv <- expsurv.simulate(alpha = 0.3, lambda = 1, N = 1000, duration = 15)
likelihood <- expsurv.likelihood(alpha = 0.3, lambda = 1, data = surv)
}
