% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialsurvival.R
\name{expsurv.simulate}
\alias{expsurv.simulate}
\title{Simulate realisations from a Parametric Exponential Mixture Model}
\usage{
expsurv.simulate(alpha, lambda, N, duration)
}
\arguments{
\item{alpha}{Proportion of individuals who experience 0 risk}

\item{lambda}{Constant hazard}

\item{N}{Number of realisations (individuals in population)}

\item{duration}{Duration of the study}
}
\value{
Data frame with 2 columns \code{time} and \code{status}, with one row for each realisation.
}
\description{
Simulate realisations from a Parametric Exponential Mixture Model
}
\examples{
surv <- expsurv.simulate(alpha = 0.3, lambda = 1, N = 1000, duration = 15)
}
